move: 등록페이지 버튼 디렉토리 명칭 파스칼케이스로 변경(배포 오류 수정)
