+++ client/src/components/NewRecipe/indexNewRecipe.ts
@@ -9,8 +9,8 @@ export { default as StepSet } from "./Steps/StepSet";
 export { default as ImgRadio } from "./Subject/ImgRadio";
 export { default as RadioBtn } from "./Subject/RadioBtn";
 export { default as TagWithBtn } from "./Tags/TagWithBtn";
-export { default as RemoveBtn } from "./Buttons/RemoveBtn";
-export { default as PlusBtn } from "./Buttons/PlusBtn";
+export { default as RemoveBtn } from "./FormButtons/RemoveBtn";
+export { default as PlusBtn } from "./FormButtons/PlusBtn";
 export { default as RequireMark } from "./RequireMark";
 export { default as SubjectFillSection } from "./Subject/SubjectFillSection";
 export { default as IngredientsFillSection } from "./Ingredients/IngredientsFillSection";
